Add files via upload
This is an arduino code for reading data from a DHT11 sensor, which is a temperature and humidity sensor.
The total time it takes to read the data from the DHT11 sensor can vary, but in the provided code, it waits for 2 seconds (2000 milliseconds) before starting the sensor, and each reading involves a series of bit-level operations. The entire process may take a few milliseconds to complete. Since the loop reads the sensor 5 times, the total time for 5 readings would be a few milliseconds multiplied by 5.